requirements.txt
discord
pillow
requests
dotenv
pyodbc